refactor: create separate components for sections
